Allow disabling mdarea globally via admin setting
