+++ js/admin.js
@@ -1,0 +1,1 @@
+export * from './src/admin';
